+++ sphinx/source/changelog.rst
@@ -12,8 +12,8 @@ Changelog
 * Added the ``--sim-load-config`` option to :ref:`tap2sna.py` (to set the value
   of a ``--sim-load`` configuration option: ``accelerator``, ``fast-load``,
   ``pause``, ``polarity``, ``timeout``, ``trace``)
-* Added the ``--tape-start`` option to :ref:`tap2sna.py` (to start the tape at
-  a specific block number)
+* Added the ``--tape-start`` and ``--tape-stop`` options to :ref:`tap2sna.py`
+  (to start or stop the tape at a specific block number)
 * Added the ``--interrupts`` option to :ref:`trace.py` (to enable the execution
   of interrupt routines)
 * :ref:`trace.py` now reads and writes the T-states counter in Z80 snapshots